// src/types/element.d.ts
import {
  I_Component_Props,
  I_Component_Emit,
  I_Component_Slot,
  I_Component_Variable,
  I_component_LifeCycle,
  I_Component_Classes,
} from './component';
import { ICS_Function } from './function';
import { INode } from './view';

export interface ICS_Element {
  id: string;
  name: string;
  key: string;
  props: Array<I_Component_Props>;
  emits: Array<I_Component_Emit>;
  slots: Array<I_Component_Slot>;
  variables: Array<I_Component_Variable>;
  nodes: Array<INode>;

  lifeCycle: Array<I_component_LifeCycle>;
  functions: Array<ICS_Function>;
  classes?: Array<I_Component_Classes>;
}